refactor(git): split repository operations
Move Git synchronization, checkout, submodule, reference, commit, worktree, URL, and error handling into focused same-package files. Split the broad Git tests along the same boundaries without changing behavior.
