+++ internal/git/latest_commit_test.go
@@ -1,0 +1,50 @@
+package git_test
+
+import (
+	"testing"
+
+	"github.com/kimdre/doco-cd/internal/config/app"
+	"github.com/kimdre/doco-cd/internal/git"
+)
+
+func TestGetLatestCommit(t *testing.T) {
+	t.Parallel()
+
+	c, err := app.GetConfig()
+	if err != nil {
+		t.Fatalf("Failed to get app config: %v", err)
+	}
+
+	url := cloneUrl
+
+	auth, err := git.GetAuthMethod(url, c.SSHPrivateKey, c.SSHPrivateKeyPassphrase, c.GitAccessToken)
+	if err != nil {
+		t.Fatalf("Failed to get auth method: %v", err)
+	}
+
+	if auth != nil {
+		t.Logf("Using auth method: %s", auth.Name())
+	} else {
+		t.Log("No auth method configured, using anonymous access")
+	}
+
+	repo, err := git.CloneRepository(t.TempDir(), url, git.MainBranch, false, c.HttpProxy, auth, c.GitCloneSubmodules, 0)
+	if err != nil {
+		t.Fatalf("Failed to clone repository: %v", err)
+	}
+
+	if repo == nil {
+		t.Fatal("Repository is nil")
+	}
+
+	commit, err := git.GetLatestCommit(repo, git.MainBranch)
+	if err != nil {
+		t.Fatalf("Failed to get latest commit: %v", err)
+	}
+
+	if commit == "" {
+		t.Fatal("Commit hash is empty")
+	}
+
+	t.Log(commit)
+}
