+++ internal/git/worktree.go
@@ -1,0 +1,90 @@
+package git
+
+import (
+	"fmt"
+	"os"
+	"path/filepath"
+	"strings"
+
+	"github.com/go-git/go-git/v5"
+
+	"github.com/kimdre/doco-cd/internal/encryption"
+)
+
+// ResetTrackedFiles resets all tracked files in the worktree To their last committed state
+// while leaving untracked files intact.
+func ResetTrackedFiles(repo *git.Repository) error {
+	worktree, err := repo.Worktree()
+	if err != nil {
+		return fmt.Errorf("failed to get worktree: %w", err)
+	}
+
+	repoRoot := worktree.Filesystem.Root()
+
+	changedFiles, err := worktree.Status()
+	if err != nil {
+		return fmt.Errorf("failed to get worktree status: %w", err)
+	}
+
+	resetFiles := make([]string, 0, len(changedFiles))
+
+	for file, status := range changedFiles {
+		// Do not touch files that are not part of the Git repository (e.g. created by a container process)
+		if status.Staging == git.Untracked {
+			continue
+		}
+
+		if shouldResetDecryptedFile(repo, repoRoot, file) {
+			resetFiles = append(resetFiles, file)
+		}
+	}
+
+	if len(resetFiles) > 0 {
+		err = worktree.Reset(&git.ResetOptions{
+			Mode:  git.HardReset,
+			Files: resetFiles,
+		})
+		if err != nil {
+			return fmt.Errorf("failed to reset worktree: %w", err)
+		}
+	}
+
+	return nil
+}
+
+// shouldResetDecryptedFile determines whether a file should be reset based on its decrypted content.
+func shouldResetDecryptedFile(repo *git.Repository, repoRoot, file string) bool {
+	headRef, err := repo.Head()
+	if err != nil {
+		return true
+	}
+
+	commit, err := repo.CommitObject(headRef.Hash())
+	if err != nil {
+		return true
+	}
+	// Get file from commit tree
+	fileObj, err := commit.File(file)
+	if err != nil {
+		return true // Not tracked, default to reset
+	}
+
+	committedBytes, err := fileObj.Contents()
+	if err != nil {
+		return true
+	}
+
+	format := encryption.GetFileFormat(fileObj.Name)
+
+	decryptedContent, err := encryption.DecryptContent([]byte(committedBytes), format)
+	if err != nil {
+		return true
+	}
+
+	workingContent, err := os.ReadFile(filepath.Join(repoRoot, file)) // #nosec G304
+	if err != nil {
+		return true
+	}
+
+	return !strings.EqualFold(string(decryptedContent), string(workingContent))
+}
